# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Huly Self-Hosted API
  version: 0.1.0
  description: |
    REST API for Huly self-hosted instances.

    ## Authentication

    All data endpoints require a workspace-scoped JWT in the `Authorization: Bearer <token>` header.
    Tokens can be minted using the CLI tool (`tools/mint-token/`) or the optional token service (`/_tokens`).

    ## Getting Started

    1. Mint a token using the CLI tool or token service
    2. Use the token to query the transactor REST API
    3. The `find-all` endpoint queries documents by class
    4. The `tx` endpoint submits transactions (create, update, delete)

  license:
    name: EPL-2.0
    url: https://www.eclipse.org/legal/epl-2.0/
servers:
  - url: https://{host}
    description: Your Huly self-hosted instance
    variables:
      host:
        default: localhost:8083
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Workspace-scoped JWT. Payload: `{ account: AccountUuid, workspace: WorkspaceUuid, exp: number }`.
        Signed with `SERVER_SECRET` using HS256.

    serverSecret:
      type: apiKey
      in: header
      name: X-Server-Secret
      description: SERVER_SECRET value — used for admin-only token service endpoints.
...
paths:
...
  /_transactor/api/v1/load-model/{workspace}:
    get:
      operationId: loadModel
      tags: [Data]
      summary: Load data model / schema
      description: |
        Returns the full data model (class hierarchy, mixins, attributes) for the workspace.
        Useful for discovering available classes and their fields.
      parameters:
        - name: workspace
          in: path
          required: true
          schema:
            type: string
          description: Workspace UUID
      responses:
        '200':
          description: Data model
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Invalid or expired token
...